remove module SimilarEntities, move functionality to EntitySearch
* add entity utility: Entity.entity_for
